backend/requirements.txt
tensorflow
opencv-python-headless
numpy
pillow
fastapi
uvicorn
python-multipart
